Use tryCatch2v, check for message property on error object, rename file
