+++ src/utility.ts
@@ -1,6 +1,7 @@
 import { outputFileSync, removeSync } from "fs-extra";
 import { tmpdir } from "os";
 import * as nodePath from "path";
+import { ErrorWithMessage } from "./types";
 
 /**
  * Create a temporary directory where tests will be executed.
@@ -21,3 +22,10 @@ export const createFile = (p: string) => outputFileSync(p, nodePath.basename(p))
  * Remove a file.
  */
 export const removeFile = (p: string) => removeSync(p);
+
+/**
+ * Check if an error has a message.
+ */
+export const isErrorMessage = (error: ErrorWithMessage | unknown): error is ErrorWithMessage =>
+  // tslint:disable-next-line: strict-type-predicates
+  (error as ErrorWithMessage).message !== undefined;
